test(server): share the raw status-line POST between the http tests
#159's oversized_post_status gains a second user in
http_transport_wiremock.rs, so move it to tests/common/raw_post.rs.
Included by #[path] rather than declared in common/mod.rs: that module
compiles into every binary saying `mod common;`, where a helper only two
of them use is dead_code under -D warnings.

Two adaptations for a caller with an ADMITTED leg, which #159 lacked.

Read only the status line, and rename accordingly. An admitted body is
answered with a chunked text/event-stream that ends itself; the
connection then idles in keep-alive, so a read to EOF waits on a close
nothing here performs.

Stop dropping the write half once the last byte is out. That half-closes
the connection, and hyper 1.11.0 builds servers with half_close(false):
with the request fully read and a response still pending the connection
is mid-message, where mid_message_detect_eof takes a client EOF on an
empty read buffer as the caller leaving and closes with nothing written
(proto/h1/conn.rs). Only an admitted body reaches that state. #159's
legs are all refused before their body is read -- the gate on the
request head, expect_json on the first frame past the cap, neither
draining what follows -- so hyper never looks for the EOF there. The
hazard therefore turns on those two library defaults, not on socket
buffer sizes; recheck them on an rmcp or hyper bump.
